fix: ens resolution only supported on mainnet (#29)
## Description

Only support ENS resolution when the user is on mainnet. Check currently
selected chain ID in Snap homepage and notify the user that ENS is only
available on mainnet. If user is not on mainnet, and inputs an ENS name,
the snap will show the error message "ENS is only supported on Ethereum
mainnet"

fixes: https://github.com/MetaMask/accounts-planning/issues/402

## Demo


https://github.com/MetaMask/snap-watch-only/assets/91970214/8e0b5111-291d-40c7-bf3e-c4e48befe239